Move VICTIM_INDEX into modelfree.env where it belongs
