+++ src/modelfree/envs/__init__.py
@@ -1,3 +1,5 @@
+import collections
+
 from gym.envs.registration import register
 from pkg_resources import resource_filename
 
@@ -34,3 +36,14 @@ register(
             'max_radius': 4.5
             },
 )
+
+# Which index does the victim play in?
+# This is really an experiment parameter rather than an environment parameter.
+# However, it's used so widely (training, evaluation, figure and video generation) and is
+# held fixed across all experiments it's convenient to treat it as if it is static.
+VICTIM_INDEX = collections.defaultdict(lambda: 0)
+VICTIM_INDEX.update({
+    # YouShallNotPass: 1 is the walker, 0 is the blocker agent.
+    # An adversarial walker makes little sense, but a blocker can be adversarial.
+    'multicomp/YouShallNotPassHumans-v0': 1,
+})
