+++ src/modelfree/visualize/training.py
@@ -9,7 +9,7 @@ from sacred import Experiment
 from sacred.observers import FileStorageObserver
 import seaborn as sns
 
-from modelfree.configs.multi.train import VICTIM_INDEX
+from modelfree.envs import VICTIM_INDEX
 from modelfree.envs.gym_compete import is_symmetric
 from modelfree.visualize import tb, util
 from modelfree.visualize.styles import STYLES
